Breaking signs/diagnostics into separate file
